% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiDA.broom.R
\name{augment_multiDA}
\alias{augment_multiDA}
\title{augment multiDA}
\usage{
augment_multiDA(x)
}
\arguments{
\item{x}{a trained \code{multiDA} object}
}
\value{
a tidy data frame, returning back the original class, matrix of features, augmented with the paritioning of each feature as given by the algorithm. In the spirit of "augment" from the broom package.
}
