% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiDA.broom.R
\name{glimpse_multiDA}
\alias{glimpse_multiDA}
\title{glimpse multiDA}
\usage{
glimpse_multiDA(x)
}
\arguments{
\item{x}{a trained \code{multiDA} object}
}
\value{
a one row data frame, with quick summaries from the algorithm. In the spirit of the "glance" function from the broom package.
}
