% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiDA.broom.R
\name{tidy_multiDA}
\alias{tidy_multiDA}
\alias{glimpse_multiDA}
\title{multiDA "broom" functions}
\usage{
tidy_multiDA(x)

glimpse_multiDA(x)
}
\arguments{
\item{x}{a trained \code{multiDA} object}

\item{x}{a trained \code{multiDA} object}
}
\value{
a tidy data frame, with key results from the trained \code{multiDA object}, namely, a data.frame of significant features and their ranks. In the spirit of "tidy" from the broom package.

a one row data frame, with quick summaries from the algorithm. In the spirit of the "glance" function from the broom package.
}
\description{
multiDA "broom" functions
}
