% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiDA.R, R/plot.multiDA.R,
%   R/predict.multiDA.R, R/print.multiDA.R
\name{multiDA}
\alias{multiDA}
\alias{plot.multiDA}
\alias{predict.multiDA}
\alias{print.multiDA}
\title{multiDA}
\usage{
multiDA(mX, vy, penalty = c("EBIC", "BIC"), equal.var = TRUE,
  set.options = c("exhaustive", "onevsrest", "onevsall", "ordinal", "user"),
  sUser = NULL)

\method{plot}{multiDA}(x, ranked = TRUE, ranks = 1:10, features = NULL)

\method{predict}{multiDA}(object, newdata, ...)

\method{print}{multiDA}(x, max.rank = 10, ...)
}
\arguments{
\item{mX}{matrix containing the training data. The rows are the sample observations, and the columns are the features.}

\item{vy}{vector of class values (for training)}

\item{penalty}{default is in the form of the EBIC, which penalises based on the number of features. If option \code{penalty="BIC"} is specified, the penalty reverts back to the BIC.}

\item{equal.var}{a \code{LOGICAL} value, indicating whether group specific variances should be equal or allowed to vary.}

\item{set.options}{options for set partition matrix S.}

\item{sUser}{if \code{set.options} is set to \code{"user"}, \code{sUser} is a user input matrix for paritions to be considered. \code{sUser} MUST be a subset of the full partition matrix..}

\item{x}{trained multiDA object}

\item{ranks}{a vector of which ranked features should be plot}

\item{object}{trained multiDA object}

\item{newdata}{matrix of observations to predict. Each row corresponds to a new observation.}

\item{...}{Any other variables which will be ignored}

\item{max.rank}{number of significant features to display. If \code{"ALL"}, all features are displayed.}

\item{x}{object to print}

\item{...}{Any other variables which will be ignored.}
}
\value{
\code{multiDA} object that contains the trained multiDA classifier

plots

list predicted class memberships of each row in newdata
}
\description{
the multiDA function

plot ranked features from a trained multiDA object

multiDA prediction of the class membership of a matrix of new observations.

Summarizes the trained multiDA classifier in a nice manner. User can select number of features to summarise
}
\examples{
#train the multiDA classifier using the SRBCT dataset, and find the resubstitution error rate

vy   <- SRBCT$vy
mX   <- SRBCT$mX
res  <- multiDA(mX, vy, equal.var=TRUE, set.options="exhaustive", penalty="EBIC")
vals <- predict(res, newdata=mX)$vy.pred          #vy.pred returns class labels
rser <- sum(vals!=vy)/length(vy)
}
