% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.multiDA.R
\name{print.multiDA}
\alias{print.multiDA}
\title{Outputs the summary for a multiDA classifier object.}
\usage{
\method{print}{multiDA}(x)
}
\arguments{
\item{x}{object to print}
}
\description{
Summarizes the trained multiDA classifier in a nice manner.
}
