% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.multiDA.R
\name{rank.multiDA}
\alias{rank.multiDA}
\title{explore ranked features from a trained multiDA object}
\usage{
rank.multiDA(x, G)
}
\arguments{
\item{x}{trained multiDA object}

\item{G}{return rankings and features based on the number of groupings (G) per feature to be considered.}

\item{...}{Any other variables which will be ignored}
}
\value{
dRank, a dataframe of summarised features, from which feature names can easily be fed into the plot.multiDA function for easy visualisation of the groupings
}
\description{
explore ranked features from a trained multiDA object
}
