% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiDA.R, R/multiDA.broom.R, R/plot.multiDA.R,
%   R/predict.multiDA.R, R/rank.multiDA.R
\name{multiDA}
\alias{multiDA}
\alias{tidy_multiDA}
\alias{plot.multiDA}
\alias{predict.multiDA}
\alias{rank.multiDA}
\title{multiDA}
\usage{
multiDA(mX, vy, penalty = c("AIC", "BIC", "GIC-2", "GIC-3", "GIC-4", "GIC-5",
  "GIC-6", "Chi-Sq", "TEST1", "TEST2", "TEST3"), equal.var = TRUE,
  set.options = c("exhaustive", "onevsrest", "onevsall", "ordinal", "user"),
  sUser = NULL)

tidy_multiDA(x)

\method{plot}{multiDA}(x, ranks = 1:10, ...)

\method{predict}{multiDA}(object, newdata, ...)

rank.multiDA(x, G)
}
\arguments{
\item{mX}{matrix containing the training data. The rows are the sample observations, and the columns are the features.}

\item{vy}{vector of class values (for training)}

\item{penalty}{choice of penalty for use in training. Options include \code{"AIC"}, \code{"BIC"}....}

\item{equal.var}{a \code{LOGICAL} value, indicating whether group specific variances should be equal or allowed to vary.}

\item{set.options}{options for set partition matrix S.}

\item{sUser}{if \code{set.options} is set to \code{"user"}, \code{sUser} is a user input matrix for paritions to be considered. \code{sUser} MUST be a subset of the full partition matrix..}

\item{x}{a trained \code{multiDA} object}

\item{ranks}{a vector of which ranked features should be plot}

\item{...}{Any other variables which will be ignored}

\item{object}{trained multiDA object}

\item{newdata}{matrix of observations to predict. Each row corresponds to a new observation.}

\item{G}{return rankings and features based on the number of groupings (G) per feature to be considered.}

\item{x}{trained multiDA object}

\item{x}{trained multiDA object}

\item{...}{Any other variables which will be ignored}
}
\value{
\code{multiDA} object that contains the trained multiDA classifier

a tidy data frame, with key results from the trained \code{multiDA object}, namely, a data.frame of significant features and their ranks. In the spirit of "tidy" from the broom package.

plots

list predicted class memberships of each row in newdata

dRank, a dataframe of summarised features, from which feature names can easily be fed into the plot.multiDA function for easy visualisation of the groupings
}
\description{
the multiDA function

multiDA "broom" functions

plot ranked features from a trained multiDA object

multiDA prediction of the class membership of a matrix of new observations.

explore ranked features from a trained multiDA object
}
\examples{
#train the multiDA classifier using the SRBCT dataset, and find the resubstitution error rate
data(SRBCT)
vy   <- SRBCT$vy
mX   <- SRBCT$mX
res  <- multiDA(mX, vy, penalty="GIC-4", equal.var=TRUE, set.options="exhaustive")
vals <- predict(res, newdata=mX)$vy.pred          #vy.pred returns class labels
rser <- sum(vals!=vy)/length(vy)
}
