% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{multiDA}
\alias{multiDA}
\title{the multiDA function}
\usage{
multiDA(x)
}
\arguments{
\item{x}{A numeric value to be squared}
}
\value{
The square of the input
}
\description{
Takes in any numeric value and square it
}
